requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Django==5.2.3
django-admin-interface==0.30.1
django-colorfield==0.14.0
django-jazzmin==3.0.1
django-nested-admin==4.1.1
idna==3.10
joblib==1.5.1
nltk==3.9.1
numpy==2.3.1
pandas==2.3.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-monkey-business==1.1.0
python-slugify==8.0.4
pytz==2025.2
qrcode==8.2
regex==2024.11.6
requests==2.32.4
scikit-learn==1.7.0
scipy==1.16.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
threadpoolctl==3.6.0
tqdm==4.67.1
tzdata==2025.2
urllib3==2.5.0
